Added notice about implementing vector space using linear. Refs #13.
This library provides an abstraction over vector spaces, but users can
still implement their own instances.

Because other vector libraries are much more common than this one, it is
helpful to make users aware that using it (e.g., if required by another
package) does not mean having to stick to less efficient vector
implementations, and they can trivially define the necessary instances
to use their preferred vector implementations underneath.

This commit adds a notice with an example of how to use it.