Adapt installation instructions to new cabal. Refs #15.
The README mentions that the library can be installed with cabal install
simple-affine-space. While true, this will not give users of current
versions of cabal access to it in their GHCi installations. It also
mentions the use of sandboxes, which have been removed in recent
versions of cabal-install.

This commit adapts the installation instructions to not use sandboxes,
and use --lib when installing simple-affine-space. That makes the
library accessible in the default environment loaded by GHCi, letting
inexperienced users run examples directly from the terminal.

[ci skip]