Move constraints from class to method signatures. Refs #17.
The constraints of Eq and Floating on VectorSpace make it impossible to
define instances that do not have a notion of equality (e.g., MSFs in
dunai).

This commit moves the constraints from the class header to the default
implementation of the methods that actually needs those constraints.

It enables users of the library to define their own implementations for
types that do not meet the constraints (e.g., not instances of Eq).