Enable tests only with GHC 8.4. Refs #21.
The dependencies of this library are minimal, yet when all tests are
enabled, it takes forever to test (consuming considerable cycles).
Documentation and hlint tests are not needed will all GHC versions.
Testing them with one version only should be enough.

This commit modifies the travis job so that tests are only installed and
executed with GHC 8.4.4.