Do not install alex or happy in CI job unless tests are enabled. Refs #22.
Alex and happy are not needed for this library, they are only needed for
the tests. Installing them in the CI job is wasteful, and is consuming
valuable resources.

This commit modifies the travis config file so that alex and happy are
not manually installed, except when the library is being tested.